+++ gaphor/plugins/diagramexport/gaphorconvert.py
