--- gaphor/tools/__init__.py
@@ -1,7 +1,0 @@
-"""
-Tools
-=====
-
-This package (``gaphor.tools``) is the place to put (command-line) utilities
-that are part of the Gaphor distribution.
-"""
