+++ pyproject.toml
@@ -61,7 +61,7 @@ isort = "^4.3.21"
 
 [tool.poetry.scripts]
 gaphor = 'gaphor.ui:main'
-gaphorconvert = 'gaphor.tools.gaphorconvert:main'
+gaphorconvert = 'gaphor.plugins.diagramexport.gaphorconvert:main'
 
 [tool.poetry.plugins."gaphor.appservices"]
 "session" = "gaphor.services.session:Session"
